requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
python-dotenv
psycopg2-binary
fastapi[standard]
pyjwt
datetime
werkzeug
asyncpg
pylint
autopep8
black
passlib
bcrypt
valkey
valkey[hiredis]